Organize api files
