+++ frigate/api/defs/query/events_query_parameters.py
