+++ frigate/api/defs/response/generic_response.py
