+++ frigate/api/review.py
@@ -12,14 +12,14 @@ from fastapi.responses import JSONResponse
 from peewee import Case, DoesNotExist, fn, operator
 from playhouse.shortcuts import model_to_dict
 
-from frigate.api.defs.generic_response import GenericResponse
-from frigate.api.defs.review_body import ReviewModifyMultipleBody
+from frigate.api.defs.response.generic_response import GenericResponse
+from frigate.api.defs.request.review_body import ReviewModifyMultipleBody
 from frigate.api.defs.review_query_parameters import (
     ReviewActivityMotionQueryParams,
     ReviewQueryParams,
     ReviewSummaryQueryParams,
 )
-from frigate.api.defs.review_responses import (
+from frigate.api.defs.response.review_response import (
     ReviewActivityMotionResponse,
     ReviewSegmentResponse,
     ReviewSummaryResponse,
@@ -364,7 +364,7 @@ def delete_reviews(body: ReviewModifyMultipleBody):
     ReviewSegment.delete().where(ReviewSegment.id << list_of_ids).execute()
 
     return JSONResponse(
-        content=({"success": True, "message": "Delete reviews"}), status_code=200
+        content=({"success": True, "message": "Deleted review items."}), status_code=200
     )
